LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-outside/values/shape-outside-ellipse-005-expected.txt


PASS test unit: cm - ellipse(50cm 100cm at 50% 50%) - inline
PASS test unit: cm - ellipse(100cm 100px at 50% 50%) - inline
PASS test unit: cm - ellipse(50cm 25% at 50% 50%) - inline
PASS test unit: cm - ellipse(25% 50cm at 50% 50%) - inline
PASS test unit: cm - ellipse(25cm closest-side at 50% 50%) - inline
PASS test unit: cm - ellipse(closest-side 75cm at 50% 50%) - inline
PASS test unit: cm - ellipse(25cm farthest-side at 50% 50%) - inline
PASS test unit: cm - ellipse(farthest-side 75cm at 50% 50%) - inline
FAIL test unit: cm - ellipse(50cm 100cm) - inline assert_equals: expected "ellipse(50cm 100cm)" but got "ellipse(50cm 100cm at 50% 50%)"
FAIL test unit: cm - ellipse(100cm 100px) - inline assert_equals: expected "ellipse(100cm 100px)" but got "ellipse(100cm 100px at 50% 50%)"
FAIL test unit: cm - ellipse(50cm 25%) - inline assert_equals: expected "ellipse(50cm 25%)" but got "ellipse(50cm 25% at 50% 50%)"
FAIL test unit: cm - ellipse(25% 50cm) - inline assert_equals: expected "ellipse(25% 50cm)" but got "ellipse(25% 50cm at 50% 50%)"
FAIL test unit: cm - ellipse(25cm closest-side) - inline assert_equals: expected "ellipse(25cm closest-side)" but got "ellipse(25cm closest-side at 50% 50%)"
FAIL test unit: cm - ellipse(closest-side 75cm) - inline assert_equals: expected "ellipse(closest-side 75cm)" but got "ellipse(closest-side 75cm at 50% 50%)"
FAIL test unit: cm - ellipse(25cm farthest-side) - inline assert_equals: expected "ellipse(25cm farthest-side)" but got "ellipse(25cm farthest-side at 50% 50%)"
FAIL test unit: cm - ellipse(farthest-side 75cm) - inline assert_equals: expected "ellipse(farthest-side 75cm)" but got "ellipse(farthest-side 75cm at 50% 50%)"
PASS test unit: mm - ellipse(50mm 100mm at 50% 50%) - inline
PASS test unit: mm - ellipse(100mm 100px at 50% 50%) - inline
PASS test unit: mm - ellipse(50mm 25% at 50% 50%) - inline
PASS test unit: mm - ellipse(25% 50mm at 50% 50%) - inline
PASS test unit: mm - ellipse(25mm closest-side at 50% 50%) - inline
PASS test unit: mm - ellipse(closest-side 75mm at 50% 50%) - inline
PASS test unit: mm - ellipse(25mm farthest-side at 50% 50%) - inline
PASS test unit: mm - ellipse(farthest-side 75mm at 50% 50%) - inline
FAIL test unit: mm - ellipse(50mm 100mm) - inline assert_equals: expected "ellipse(50mm 100mm)" but got "ellipse(50mm 100mm at 50% 50%)"
FAIL test unit: mm - ellipse(100mm 100px) - inline assert_equals: expected "ellipse(100mm 100px)" but got "ellipse(100mm 100px at 50% 50%)"
FAIL test unit: mm - ellipse(50mm 25%) - inline assert_equals: expected "ellipse(50mm 25%)" but got "ellipse(50mm 25% at 50% 50%)"
FAIL test unit: mm - ellipse(25% 50mm) - inline assert_equals: expected "ellipse(25% 50mm)" but got "ellipse(25% 50mm at 50% 50%)"
FAIL test unit: mm - ellipse(25mm closest-side) - inline assert_equals: expected "ellipse(25mm closest-side)" but got "ellipse(25mm closest-side at 50% 50%)"
FAIL test unit: mm - ellipse(closest-side 75mm) - inline assert_equals: expected "ellipse(closest-side 75mm)" but got "ellipse(closest-side 75mm at 50% 50%)"
FAIL test unit: mm - ellipse(25mm farthest-side) - inline assert_equals: expected "ellipse(25mm farthest-side)" but got "ellipse(25mm farthest-side at 50% 50%)"
FAIL test unit: mm - ellipse(farthest-side 75mm) - inline assert_equals: expected "ellipse(farthest-side 75mm)" but got "ellipse(farthest-side 75mm at 50% 50%)"
PASS test unit: in - ellipse(50in 100in at 50% 50%) - inline
PASS test unit: in - ellipse(100in 100px at 50% 50%) - inline
PASS test unit: in - ellipse(50in 25% at 50% 50%) - inline
PASS test unit: in - ellipse(25% 50in at 50% 50%) - inline
PASS test unit: in - ellipse(25in closest-side at 50% 50%) - inline
PASS test unit: in - ellipse(closest-side 75in at 50% 50%) - inline
PASS test unit: in - ellipse(25in farthest-side at 50% 50%) - inline
PASS test unit: in - ellipse(farthest-side 75in at 50% 50%) - inline
FAIL test unit: in - ellipse(50in 100in) - inline assert_equals: expected "ellipse(50in 100in)" but got "ellipse(50in 100in at 50% 50%)"
FAIL test unit: in - ellipse(100in 100px) - inline assert_equals: expected "ellipse(100in 100px)" but got "ellipse(100in 100px at 50% 50%)"
FAIL test unit: in - ellipse(50in 25%) - inline assert_equals: expected "ellipse(50in 25%)" but got "ellipse(50in 25% at 50% 50%)"
FAIL test unit: in - ellipse(25% 50in) - inline assert_equals: expected "ellipse(25% 50in)" but got "ellipse(25% 50in at 50% 50%)"
FAIL test unit: in - ellipse(25in closest-side) - inline assert_equals: expected "ellipse(25in closest-side)" but got "ellipse(25in closest-side at 50% 50%)"
FAIL test unit: in - ellipse(closest-side 75in) - inline assert_equals: expected "ellipse(closest-side 75in)" but got "ellipse(closest-side 75in at 50% 50%)"
FAIL test unit: in - ellipse(25in farthest-side) - inline assert_equals: expected "ellipse(25in farthest-side)" but got "ellipse(25in farthest-side at 50% 50%)"
FAIL test unit: in - ellipse(farthest-side 75in) - inline assert_equals: expected "ellipse(farthest-side 75in)" but got "ellipse(farthest-side 75in at 50% 50%)"
PASS test unit: pt - ellipse(50pt 100pt at 50% 50%) - inline
PASS test unit: pt - ellipse(100pt 100px at 50% 50%) - inline
PASS test unit: pt - ellipse(50pt 25% at 50% 50%) - inline
PASS test unit: pt - ellipse(25% 50pt at 50% 50%) - inline
PASS test unit: pt - ellipse(25pt closest-side at 50% 50%) - inline
PASS test unit: pt - ellipse(closest-side 75pt at 50% 50%) - inline
PASS test unit: pt - ellipse(25pt farthest-side at 50% 50%) - inline
PASS test unit: pt - ellipse(farthest-side 75pt at 50% 50%) - inline
FAIL test unit: pt - ellipse(50pt 100pt) - inline assert_equals: expected "ellipse(50pt 100pt)" but got "ellipse(50pt 100pt at 50% 50%)"
FAIL test unit: pt - ellipse(100pt 100px) - inline assert_equals: expected "ellipse(100pt 100px)" but got "ellipse(100pt 100px at 50% 50%)"
FAIL test unit: pt - ellipse(50pt 25%) - inline assert_equals: expected "ellipse(50pt 25%)" but got "ellipse(50pt 25% at 50% 50%)"
FAIL test unit: pt - ellipse(25% 50pt) - inline assert_equals: expected "ellipse(25% 50pt)" but got "ellipse(25% 50pt at 50% 50%)"
FAIL test unit: pt - ellipse(25pt closest-side) - inline assert_equals: expected "ellipse(25pt closest-side)" but got "ellipse(25pt closest-side at 50% 50%)"
FAIL test unit: pt - ellipse(closest-side 75pt) - inline assert_equals: expected "ellipse(closest-side 75pt)" but got "ellipse(closest-side 75pt at 50% 50%)"
FAIL test unit: pt - ellipse(25pt farthest-side) - inline assert_equals: expected "ellipse(25pt farthest-side)" but got "ellipse(25pt farthest-side at 50% 50%)"
FAIL test unit: pt - ellipse(farthest-side 75pt) - inline assert_equals: expected "ellipse(farthest-side 75pt)" but got "ellipse(farthest-side 75pt at 50% 50%)"
PASS test unit: pc - ellipse(50pc 100pc at 50% 50%) - inline
PASS test unit: pc - ellipse(100pc 100px at 50% 50%) - inline
PASS test unit: pc - ellipse(50pc 25% at 50% 50%) - inline
PASS test unit: pc - ellipse(25% 50pc at 50% 50%) - inline
PASS test unit: pc - ellipse(25pc closest-side at 50% 50%) - inline
PASS test unit: pc - ellipse(closest-side 75pc at 50% 50%) - inline
PASS test unit: pc - ellipse(25pc farthest-side at 50% 50%) - inline
PASS test unit: pc - ellipse(farthest-side 75pc at 50% 50%) - inline
FAIL test unit: pc - ellipse(50pc 100pc) - inline assert_equals: expected "ellipse(50pc 100pc)" but got "ellipse(50pc 100pc at 50% 50%)"
FAIL test unit: pc - ellipse(100pc 100px) - inline assert_equals: expected "ellipse(100pc 100px)" but got "ellipse(100pc 100px at 50% 50%)"
FAIL test unit: pc - ellipse(50pc 25%) - inline assert_equals: expected "ellipse(50pc 25%)" but got "ellipse(50pc 25% at 50% 50%)"
FAIL test unit: pc - ellipse(25% 50pc) - inline assert_equals: expected "ellipse(25% 50pc)" but got "ellipse(25% 50pc at 50% 50%)"
FAIL test unit: pc - ellipse(25pc closest-side) - inline assert_equals: expected "ellipse(25pc closest-side)" but got "ellipse(25pc closest-side at 50% 50%)"
FAIL test unit: pc - ellipse(closest-side 75pc) - inline assert_equals: expected "ellipse(closest-side 75pc)" but got "ellipse(closest-side 75pc at 50% 50%)"
FAIL test unit: pc - ellipse(25pc farthest-side) - inline assert_equals: expected "ellipse(25pc farthest-side)" but got "ellipse(25pc farthest-side at 50% 50%)"
FAIL test unit: pc - ellipse(farthest-side 75pc) - inline assert_equals: expected "ellipse(farthest-side 75pc)" but got "ellipse(farthest-side 75pc at 50% 50%)"
PASS test unit: em - ellipse(50em 100em at 50% 50%) - inline
PASS test unit: em - ellipse(100em 100px at 50% 50%) - inline
PASS test unit: em - ellipse(50em 25% at 50% 50%) - inline
PASS test unit: em - ellipse(25% 50em at 50% 50%) - inline
PASS test unit: em - ellipse(25em closest-side at 50% 50%) - inline
PASS test unit: em - ellipse(closest-side 75em at 50% 50%) - inline
PASS test unit: em - ellipse(25em farthest-side at 50% 50%) - inline
PASS test unit: em - ellipse(farthest-side 75em at 50% 50%) - inline
FAIL test unit: em - ellipse(50em 100em) - inline assert_equals: expected "ellipse(50em 100em)" but got "ellipse(50em 100em at 50% 50%)"
FAIL test unit: em - ellipse(100em 100px) - inline assert_equals: expected "ellipse(100em 100px)" but got "ellipse(100em 100px at 50% 50%)"
FAIL test unit: em - ellipse(50em 25%) - inline assert_equals: expected "ellipse(50em 25%)" but got "ellipse(50em 25% at 50% 50%)"
FAIL test unit: em - ellipse(25% 50em) - inline assert_equals: expected "ellipse(25% 50em)" but got "ellipse(25% 50em at 50% 50%)"
FAIL test unit: em - ellipse(25em closest-side) - inline assert_equals: expected "ellipse(25em closest-side)" but got "ellipse(25em closest-side at 50% 50%)"
FAIL test unit: em - ellipse(closest-side 75em) - inline assert_equals: expected "ellipse(closest-side 75em)" but got "ellipse(closest-side 75em at 50% 50%)"
FAIL test unit: em - ellipse(25em farthest-side) - inline assert_equals: expected "ellipse(25em farthest-side)" but got "ellipse(25em farthest-side at 50% 50%)"
FAIL test unit: em - ellipse(farthest-side 75em) - inline assert_equals: expected "ellipse(farthest-side 75em)" but got "ellipse(farthest-side 75em at 50% 50%)"
PASS test unit: ex - ellipse(50ex 100ex at 50% 50%) - inline
PASS test unit: ex - ellipse(100ex 100px at 50% 50%) - inline
PASS test unit: ex - ellipse(50ex 25% at 50% 50%) - inline
PASS test unit: ex - ellipse(25% 50ex at 50% 50%) - inline
PASS test unit: ex - ellipse(25ex closest-side at 50% 50%) - inline
PASS test unit: ex - ellipse(closest-side 75ex at 50% 50%) - inline
PASS test unit: ex - ellipse(25ex farthest-side at 50% 50%) - inline
PASS test unit: ex - ellipse(farthest-side 75ex at 50% 50%) - inline
FAIL test unit: ex - ellipse(50ex 100ex) - inline assert_equals: expected "ellipse(50ex 100ex)" but got "ellipse(50ex 100ex at 50% 50%)"
FAIL test unit: ex - ellipse(100ex 100px) - inline assert_equals: expected "ellipse(100ex 100px)" but got "ellipse(100ex 100px at 50% 50%)"
FAIL test unit: ex - ellipse(50ex 25%) - inline assert_equals: expected "ellipse(50ex 25%)" but got "ellipse(50ex 25% at 50% 50%)"
FAIL test unit: ex - ellipse(25% 50ex) - inline assert_equals: expected "ellipse(25% 50ex)" but got "ellipse(25% 50ex at 50% 50%)"
FAIL test unit: ex - ellipse(25ex closest-side) - inline assert_equals: expected "ellipse(25ex closest-side)" but got "ellipse(25ex closest-side at 50% 50%)"
FAIL test unit: ex - ellipse(closest-side 75ex) - inline assert_equals: expected "ellipse(closest-side 75ex)" but got "ellipse(closest-side 75ex at 50% 50%)"
FAIL test unit: ex - ellipse(25ex farthest-side) - inline assert_equals: expected "ellipse(25ex farthest-side)" but got "ellipse(25ex farthest-side at 50% 50%)"
FAIL test unit: ex - ellipse(farthest-side 75ex) - inline assert_equals: expected "ellipse(farthest-side 75ex)" but got "ellipse(farthest-side 75ex at 50% 50%)"
PASS test unit: ch - ellipse(50ch 100ch at 50% 50%) - inline
PASS test unit: ch - ellipse(100ch 100px at 50% 50%) - inline
PASS test unit: ch - ellipse(50ch 25% at 50% 50%) - inline
PASS test unit: ch - ellipse(25% 50ch at 50% 50%) - inline
PASS test unit: ch - ellipse(25ch closest-side at 50% 50%) - inline
PASS test unit: ch - ellipse(closest-side 75ch at 50% 50%) - inline
PASS test unit: ch - ellipse(25ch farthest-side at 50% 50%) - inline
PASS test unit: ch - ellipse(farthest-side 75ch at 50% 50%) - inline
FAIL test unit: ch - ellipse(50ch 100ch) - inline assert_equals: expected "ellipse(50ch 100ch)" but got "ellipse(50ch 100ch at 50% 50%)"
FAIL test unit: ch - ellipse(100ch 100px) - inline assert_equals: expected "ellipse(100ch 100px)" but got "ellipse(100ch 100px at 50% 50%)"
FAIL test unit: ch - ellipse(50ch 25%) - inline assert_equals: expected "ellipse(50ch 25%)" but got "ellipse(50ch 25% at 50% 50%)"
FAIL test unit: ch - ellipse(25% 50ch) - inline assert_equals: expected "ellipse(25% 50ch)" but got "ellipse(25% 50ch at 50% 50%)"
FAIL test unit: ch - ellipse(25ch closest-side) - inline assert_equals: expected "ellipse(25ch closest-side)" but got "ellipse(25ch closest-side at 50% 50%)"
FAIL test unit: ch - ellipse(closest-side 75ch) - inline assert_equals: expected "ellipse(closest-side 75ch)" but got "ellipse(closest-side 75ch at 50% 50%)"
FAIL test unit: ch - ellipse(25ch farthest-side) - inline assert_equals: expected "ellipse(25ch farthest-side)" but got "ellipse(25ch farthest-side at 50% 50%)"
FAIL test unit: ch - ellipse(farthest-side 75ch) - inline assert_equals: expected "ellipse(farthest-side 75ch)" but got "ellipse(farthest-side 75ch at 50% 50%)"
PASS test unit: rem - ellipse(50rem 100rem at 50% 50%) - inline
PASS test unit: rem - ellipse(100rem 100px at 50% 50%) - inline
PASS test unit: rem - ellipse(50rem 25% at 50% 50%) - inline
PASS test unit: rem - ellipse(25% 50rem at 50% 50%) - inline
PASS test unit: rem - ellipse(25rem closest-side at 50% 50%) - inline
PASS test unit: rem - ellipse(closest-side 75rem at 50% 50%) - inline
PASS test unit: rem - ellipse(25rem farthest-side at 50% 50%) - inline
PASS test unit: rem - ellipse(farthest-side 75rem at 50% 50%) - inline
FAIL test unit: rem - ellipse(50rem 100rem) - inline assert_equals: expected "ellipse(50rem 100rem)" but got "ellipse(50rem 100rem at 50% 50%)"
FAIL test unit: rem - ellipse(100rem 100px) - inline assert_equals: expected "ellipse(100rem 100px)" but got "ellipse(100rem 100px at 50% 50%)"
FAIL test unit: rem - ellipse(50rem 25%) - inline assert_equals: expected "ellipse(50rem 25%)" but got "ellipse(50rem 25% at 50% 50%)"
FAIL test unit: rem - ellipse(25% 50rem) - inline assert_equals: expected "ellipse(25% 50rem)" but got "ellipse(25% 50rem at 50% 50%)"
FAIL test unit: rem - ellipse(25rem closest-side) - inline assert_equals: expected "ellipse(25rem closest-side)" but got "ellipse(25rem closest-side at 50% 50%)"
FAIL test unit: rem - ellipse(closest-side 75rem) - inline assert_equals: expected "ellipse(closest-side 75rem)" but got "ellipse(closest-side 75rem at 50% 50%)"
FAIL test unit: rem - ellipse(25rem farthest-side) - inline assert_equals: expected "ellipse(25rem farthest-side)" but got "ellipse(25rem farthest-side at 50% 50%)"
FAIL test unit: rem - ellipse(farthest-side 75rem) - inline assert_equals: expected "ellipse(farthest-side 75rem)" but got "ellipse(farthest-side 75rem at 50% 50%)"
PASS test unit: vw - ellipse(50vw 100vw at 50% 50%) - inline
PASS test unit: vw - ellipse(100vw 100px at 50% 50%) - inline
PASS test unit: vw - ellipse(50vw 25% at 50% 50%) - inline
PASS test unit: vw - ellipse(25% 50vw at 50% 50%) - inline
PASS test unit: vw - ellipse(25vw closest-side at 50% 50%) - inline
PASS test unit: vw - ellipse(closest-side 75vw at 50% 50%) - inline
PASS test unit: vw - ellipse(25vw farthest-side at 50% 50%) - inline
PASS test unit: vw - ellipse(farthest-side 75vw at 50% 50%) - inline
FAIL test unit: vw - ellipse(50vw 100vw) - inline assert_equals: expected "ellipse(50vw 100vw)" but got "ellipse(50vw 100vw at 50% 50%)"
FAIL test unit: vw - ellipse(100vw 100px) - inline assert_equals: expected "ellipse(100vw 100px)" but got "ellipse(100vw 100px at 50% 50%)"
FAIL test unit: vw - ellipse(50vw 25%) - inline assert_equals: expected "ellipse(50vw 25%)" but got "ellipse(50vw 25% at 50% 50%)"
FAIL test unit: vw - ellipse(25% 50vw) - inline assert_equals: expected "ellipse(25% 50vw)" but got "ellipse(25% 50vw at 50% 50%)"
FAIL test unit: vw - ellipse(25vw closest-side) - inline assert_equals: expected "ellipse(25vw closest-side)" but got "ellipse(25vw closest-side at 50% 50%)"
FAIL test unit: vw - ellipse(closest-side 75vw) - inline assert_equals: expected "ellipse(closest-side 75vw)" but got "ellipse(closest-side 75vw at 50% 50%)"
FAIL test unit: vw - ellipse(25vw farthest-side) - inline assert_equals: expected "ellipse(25vw farthest-side)" but got "ellipse(25vw farthest-side at 50% 50%)"
FAIL test unit: vw - ellipse(farthest-side 75vw) - inline assert_equals: expected "ellipse(farthest-side 75vw)" but got "ellipse(farthest-side 75vw at 50% 50%)"
PASS test unit: vh - ellipse(50vh 100vh at 50% 50%) - inline
PASS test unit: vh - ellipse(100vh 100px at 50% 50%) - inline
PASS test unit: vh - ellipse(50vh 25% at 50% 50%) - inline
PASS test unit: vh - ellipse(25% 50vh at 50% 50%) - inline
PASS test unit: vh - ellipse(25vh closest-side at 50% 50%) - inline
PASS test unit: vh - ellipse(closest-side 75vh at 50% 50%) - inline
PASS test unit: vh - ellipse(25vh farthest-side at 50% 50%) - inline
PASS test unit: vh - ellipse(farthest-side 75vh at 50% 50%) - inline
FAIL test unit: vh - ellipse(50vh 100vh) - inline assert_equals: expected "ellipse(50vh 100vh)" but got "ellipse(50vh 100vh at 50% 50%)"
FAIL test unit: vh - ellipse(100vh 100px) - inline assert_equals: expected "ellipse(100vh 100px)" but got "ellipse(100vh 100px at 50% 50%)"
FAIL test unit: vh - ellipse(50vh 25%) - inline assert_equals: expected "ellipse(50vh 25%)" but got "ellipse(50vh 25% at 50% 50%)"
FAIL test unit: vh - ellipse(25% 50vh) - inline assert_equals: expected "ellipse(25% 50vh)" but got "ellipse(25% 50vh at 50% 50%)"
FAIL test unit: vh - ellipse(25vh closest-side) - inline assert_equals: expected "ellipse(25vh closest-side)" but got "ellipse(25vh closest-side at 50% 50%)"
FAIL test unit: vh - ellipse(closest-side 75vh) - inline assert_equals: expected "ellipse(closest-side 75vh)" but got "ellipse(closest-side 75vh at 50% 50%)"
FAIL test unit: vh - ellipse(25vh farthest-side) - inline assert_equals: expected "ellipse(25vh farthest-side)" but got "ellipse(25vh farthest-side at 50% 50%)"
FAIL test unit: vh - ellipse(farthest-side 75vh) - inline assert_equals: expected "ellipse(farthest-side 75vh)" but got "ellipse(farthest-side 75vh at 50% 50%)"
PASS test unit: vmin - ellipse(50vmin 100vmin at 50% 50%) - inline
PASS test unit: vmin - ellipse(100vmin 100px at 50% 50%) - inline
PASS test unit: vmin - ellipse(50vmin 25% at 50% 50%) - inline
PASS test unit: vmin - ellipse(25% 50vmin at 50% 50%) - inline
PASS test unit: vmin - ellipse(25vmin closest-side at 50% 50%) - inline
PASS test unit: vmin - ellipse(closest-side 75vmin at 50% 50%) - inline
PASS test unit: vmin - ellipse(25vmin farthest-side at 50% 50%) - inline
PASS test unit: vmin - ellipse(farthest-side 75vmin at 50% 50%) - inline
FAIL test unit: vmin - ellipse(50vmin 100vmin) - inline assert_equals: expected "ellipse(50vmin 100vmin)" but got "ellipse(50vmin 100vmin at 50% 50%)"
FAIL test unit: vmin - ellipse(100vmin 100px) - inline assert_equals: expected "ellipse(100vmin 100px)" but got "ellipse(100vmin 100px at 50% 50%)"
FAIL test unit: vmin - ellipse(50vmin 25%) - inline assert_equals: expected "ellipse(50vmin 25%)" but got "ellipse(50vmin 25% at 50% 50%)"
FAIL test unit: vmin - ellipse(25% 50vmin) - inline assert_equals: expected "ellipse(25% 50vmin)" but got "ellipse(25% 50vmin at 50% 50%)"
FAIL test unit: vmin - ellipse(25vmin closest-side) - inline assert_equals: expected "ellipse(25vmin closest-side)" but got "ellipse(25vmin closest-side at 50% 50%)"
FAIL test unit: vmin - ellipse(closest-side 75vmin) - inline assert_equals: expected "ellipse(closest-side 75vmin)" but got "ellipse(closest-side 75vmin at 50% 50%)"
FAIL test unit: vmin - ellipse(25vmin farthest-side) - inline assert_equals: expected "ellipse(25vmin farthest-side)" but got "ellipse(25vmin farthest-side at 50% 50%)"
FAIL test unit: vmin - ellipse(farthest-side 75vmin) - inline assert_equals: expected "ellipse(farthest-side 75vmin)" but got "ellipse(farthest-side 75vmin at 50% 50%)"
PASS test unit: vmax - ellipse(50vmax 100vmax at 50% 50%) - inline
PASS test unit: vmax - ellipse(100vmax 100px at 50% 50%) - inline
PASS test unit: vmax - ellipse(50vmax 25% at 50% 50%) - inline
PASS test unit: vmax - ellipse(25% 50vmax at 50% 50%) - inline
PASS test unit: vmax - ellipse(25vmax closest-side at 50% 50%) - inline
PASS test unit: vmax - ellipse(closest-side 75vmax at 50% 50%) - inline
PASS test unit: vmax - ellipse(25vmax farthest-side at 50% 50%) - inline
PASS test unit: vmax - ellipse(farthest-side 75vmax at 50% 50%) - inline
FAIL test unit: vmax - ellipse(50vmax 100vmax) - inline assert_equals: expected "ellipse(50vmax 100vmax)" but got "ellipse(50vmax 100vmax at 50% 50%)"
FAIL test unit: vmax - ellipse(100vmax 100px) - inline assert_equals: expected "ellipse(100vmax 100px)" but got "ellipse(100vmax 100px at 50% 50%)"
FAIL test unit: vmax - ellipse(50vmax 25%) - inline assert_equals: expected "ellipse(50vmax 25%)" but got "ellipse(50vmax 25% at 50% 50%)"
FAIL test unit: vmax - ellipse(25% 50vmax) - inline assert_equals: expected "ellipse(25% 50vmax)" but got "ellipse(25% 50vmax at 50% 50%)"
FAIL test unit: vmax - ellipse(25vmax closest-side) - inline assert_equals: expected "ellipse(25vmax closest-side)" but got "ellipse(25vmax closest-side at 50% 50%)"
FAIL test unit: vmax - ellipse(closest-side 75vmax) - inline assert_equals: expected "ellipse(closest-side 75vmax)" but got "ellipse(closest-side 75vmax at 50% 50%)"
FAIL test unit: vmax - ellipse(25vmax farthest-side) - inline assert_equals: expected "ellipse(25vmax farthest-side)" but got "ellipse(25vmax farthest-side at 50% 50%)"
FAIL test unit: vmax - ellipse(farthest-side 75vmax) - inline assert_equals: expected "ellipse(farthest-side 75vmax)" but got "ellipse(farthest-side 75vmax at 50% 50%)"
PASS test unit: cm - ellipse(50cm 100cm at 50% 50%) - computed
PASS test unit: cm - ellipse(100cm 100px at 50% 50%) - computed
PASS test unit: cm - ellipse(50cm 25% at 50% 50%) - computed
PASS test unit: cm - ellipse(25% 50cm at 50% 50%) - computed
PASS test unit: cm - ellipse(25cm closest-side at 50% 50%) - computed
PASS test unit: cm - ellipse(closest-side 75cm at 50% 50%) - computed
PASS test unit: cm - ellipse(25cm farthest-side at 50% 50%) - computed
PASS test unit: cm - ellipse(farthest-side 75cm at 50% 50%) - computed
FAIL test unit: cm - ellipse(50cm 100cm) - computed assert_equals: expected "ellipse(1889.76px 3779.53px)" but got "ellipse(1889.76px 3779.53px at 50% 50%)"
FAIL test unit: cm - ellipse(100cm 100px) - computed assert_equals: expected "ellipse(3779.53px 100px)" but got "ellipse(3779.53px 100px at 50% 50%)"
FAIL test unit: cm - ellipse(50cm 25%) - computed assert_equals: expected "ellipse(1889.76px 25%)" but got "ellipse(1889.76px 25% at 50% 50%)"
FAIL test unit: cm - ellipse(25% 50cm) - computed assert_equals: expected "ellipse(25% 1889.76px)" but got "ellipse(25% 1889.76px at 50% 50%)"
FAIL test unit: cm - ellipse(25cm closest-side) - computed assert_equals: expected "ellipse(944.882px closest-side)" but got "ellipse(944.882px closest-side at 50% 50%)"
FAIL test unit: cm - ellipse(closest-side 75cm) - computed assert_equals: expected "ellipse(closest-side 2834.65px)" but got "ellipse(closest-side 2834.65px at 50% 50%)"
FAIL test unit: cm - ellipse(25cm farthest-side) - computed assert_equals: expected "ellipse(944.882px farthest-side)" but got "ellipse(944.882px farthest-side at 50% 50%)"
FAIL test unit: cm - ellipse(farthest-side 75cm) - computed assert_equals: expected "ellipse(farthest-side 2834.65px)" but got "ellipse(farthest-side 2834.65px at 50% 50%)"
PASS test unit: mm - ellipse(50mm 100mm at 50% 50%) - computed
PASS test unit: mm - ellipse(100mm 100px at 50% 50%) - computed
PASS test unit: mm - ellipse(50mm 25% at 50% 50%) - computed
PASS test unit: mm - ellipse(25% 50mm at 50% 50%) - computed
PASS test unit: mm - ellipse(25mm closest-side at 50% 50%) - computed
PASS test unit: mm - ellipse(closest-side 75mm at 50% 50%) - computed
PASS test unit: mm - ellipse(25mm farthest-side at 50% 50%) - computed
PASS test unit: mm - ellipse(farthest-side 75mm at 50% 50%) - computed
FAIL test unit: mm - ellipse(50mm 100mm) - computed assert_equals: expected "ellipse(188.976px 377.953px)" but got "ellipse(188.976px 377.953px at 50% 50%)"
FAIL test unit: mm - ellipse(100mm 100px) - computed assert_equals: expected "ellipse(377.953px 100px)" but got "ellipse(377.953px 100px at 50% 50%)"
FAIL test unit: mm - ellipse(50mm 25%) - computed assert_equals: expected "ellipse(188.976px 25%)" but got "ellipse(188.976px 25% at 50% 50%)"
FAIL test unit: mm - ellipse(25% 50mm) - computed assert_equals: expected "ellipse(25% 188.976px)" but got "ellipse(25% 188.976px at 50% 50%)"
FAIL test unit: mm - ellipse(25mm closest-side) - computed assert_equals: expected "ellipse(94.4882px closest-side)" but got "ellipse(94.4882px closest-side at 50% 50%)"
FAIL test unit: mm - ellipse(closest-side 75mm) - computed assert_equals: expected "ellipse(closest-side 283.465px)" but got "ellipse(closest-side 283.465px at 50% 50%)"
FAIL test unit: mm - ellipse(25mm farthest-side) - computed assert_equals: expected "ellipse(94.4882px farthest-side)" but got "ellipse(94.4882px farthest-side at 50% 50%)"
FAIL test unit: mm - ellipse(farthest-side 75mm) - computed assert_equals: expected "ellipse(farthest-side 283.465px)" but got "ellipse(farthest-side 283.465px at 50% 50%)"
PASS test unit: in - ellipse(50in 100in at 50% 50%) - computed
PASS test unit: in - ellipse(100in 100px at 50% 50%) - computed
PASS test unit: in - ellipse(50in 25% at 50% 50%) - computed
PASS test unit: in - ellipse(25% 50in at 50% 50%) - computed
PASS test unit: in - ellipse(25in closest-side at 50% 50%) - computed
PASS test unit: in - ellipse(closest-side 75in at 50% 50%) - computed
PASS test unit: in - ellipse(25in farthest-side at 50% 50%) - computed
PASS test unit: in - ellipse(farthest-side 75in at 50% 50%) - computed
FAIL test unit: in - ellipse(50in 100in) - computed assert_equals: expected "ellipse(4800px 9600px)" but got "ellipse(4800px 9600px at 50% 50%)"
FAIL test unit: in - ellipse(100in 100px) - computed assert_equals: expected "ellipse(9600px 100px)" but got "ellipse(9600px 100px at 50% 50%)"
FAIL test unit: in - ellipse(50in 25%) - computed assert_equals: expected "ellipse(4800px 25%)" but got "ellipse(4800px 25% at 50% 50%)"
FAIL test unit: in - ellipse(25% 50in) - computed assert_equals: expected "ellipse(25% 4800px)" but got "ellipse(25% 4800px at 50% 50%)"
FAIL test unit: in - ellipse(25in closest-side) - computed assert_equals: expected "ellipse(2400px closest-side)" but got "ellipse(2400px closest-side at 50% 50%)"
FAIL test unit: in - ellipse(closest-side 75in) - computed assert_equals: expected "ellipse(closest-side 7200px)" but got "ellipse(closest-side 7200px at 50% 50%)"
FAIL test unit: in - ellipse(25in farthest-side) - computed assert_equals: expected "ellipse(2400px farthest-side)" but got "ellipse(2400px farthest-side at 50% 50%)"
FAIL test unit: in - ellipse(farthest-side 75in) - computed assert_equals: expected "ellipse(farthest-side 7200px)" but got "ellipse(farthest-side 7200px at 50% 50%)"
PASS test unit: pt - ellipse(50pt 100pt at 50% 50%) - computed
PASS test unit: pt - ellipse(100pt 100px at 50% 50%) - computed
PASS test unit: pt - ellipse(50pt 25% at 50% 50%) - computed
PASS test unit: pt - ellipse(25% 50pt at 50% 50%) - computed
PASS test unit: pt - ellipse(25pt closest-side at 50% 50%) - computed
PASS test unit: pt - ellipse(closest-side 75pt at 50% 50%) - computed
PASS test unit: pt - ellipse(25pt farthest-side at 50% 50%) - computed
PASS test unit: pt - ellipse(farthest-side 75pt at 50% 50%) - computed
FAIL test unit: pt - ellipse(50pt 100pt) - computed assert_equals: expected "ellipse(66.6667px 133.333px)" but got "ellipse(66.6667px 133.333px at 50% 50%)"
FAIL test unit: pt - ellipse(100pt 100px) - computed assert_equals: expected "ellipse(133.333px 100px)" but got "ellipse(133.333px 100px at 50% 50%)"
FAIL test unit: pt - ellipse(50pt 25%) - computed assert_equals: expected "ellipse(66.6667px 25%)" but got "ellipse(66.6667px 25% at 50% 50%)"
FAIL test unit: pt - ellipse(25% 50pt) - computed assert_equals: expected "ellipse(25% 66.6667px)" but got "ellipse(25% 66.6667px at 50% 50%)"
FAIL test unit: pt - ellipse(25pt closest-side) - computed assert_equals: expected "ellipse(33.3333px closest-side)" but got "ellipse(33.3333px closest-side at 50% 50%)"
FAIL test unit: pt - ellipse(closest-side 75pt) - computed assert_equals: expected "ellipse(closest-side 100px)" but got "ellipse(closest-side 100px at 50% 50%)"
FAIL test unit: pt - ellipse(25pt farthest-side) - computed assert_equals: expected "ellipse(33.3333px farthest-side)" but got "ellipse(33.3333px farthest-side at 50% 50%)"
FAIL test unit: pt - ellipse(farthest-side 75pt) - computed assert_equals: expected "ellipse(farthest-side 100px)" but got "ellipse(farthest-side 100px at 50% 50%)"
PASS test unit: pc - ellipse(50pc 100pc at 50% 50%) - computed
PASS test unit: pc - ellipse(100pc 100px at 50% 50%) - computed
PASS test unit: pc - ellipse(50pc 25% at 50% 50%) - computed
PASS test unit: pc - ellipse(25% 50pc at 50% 50%) - computed
PASS test unit: pc - ellipse(25pc closest-side at 50% 50%) - computed
PASS test unit: pc - ellipse(closest-side 75pc at 50% 50%) - computed
PASS test unit: pc - ellipse(25pc farthest-side at 50% 50%) - computed
PASS test unit: pc - ellipse(farthest-side 75pc at 50% 50%) - computed
FAIL test unit: pc - ellipse(50pc 100pc) - computed assert_equals: expected "ellipse(800px 1600px)" but got "ellipse(800px 1600px at 50% 50%)"
FAIL test unit: pc - ellipse(100pc 100px) - computed assert_equals: expected "ellipse(1600px 100px)" but got "ellipse(1600px 100px at 50% 50%)"
FAIL test unit: pc - ellipse(50pc 25%) - computed assert_equals: expected "ellipse(800px 25%)" but got "ellipse(800px 25% at 50% 50%)"
FAIL test unit: pc - ellipse(25% 50pc) - computed assert_equals: expected "ellipse(25% 800px)" but got "ellipse(25% 800px at 50% 50%)"
FAIL test unit: pc - ellipse(25pc closest-side) - computed assert_equals: expected "ellipse(400px closest-side)" but got "ellipse(400px closest-side at 50% 50%)"
FAIL test unit: pc - ellipse(closest-side 75pc) - computed assert_equals: expected "ellipse(closest-side 1200px)" but got "ellipse(closest-side 1200px at 50% 50%)"
FAIL test unit: pc - ellipse(25pc farthest-side) - computed assert_equals: expected "ellipse(400px farthest-side)" but got "ellipse(400px farthest-side at 50% 50%)"
FAIL test unit: pc - ellipse(farthest-side 75pc) - computed assert_equals: expected "ellipse(farthest-side 1200px)" but got "ellipse(farthest-side 1200px at 50% 50%)"
PASS test unit: em - ellipse(50em 100em at 50% 50%) - computed
PASS test unit: em - ellipse(100em 100px at 50% 50%) - computed
PASS test unit: em - ellipse(50em 25% at 50% 50%) - computed
PASS test unit: em - ellipse(25% 50em at 50% 50%) - computed
PASS test unit: em - ellipse(25em closest-side at 50% 50%) - computed
PASS test unit: em - ellipse(closest-side 75em at 50% 50%) - computed
PASS test unit: em - ellipse(25em farthest-side at 50% 50%) - computed
PASS test unit: em - ellipse(farthest-side 75em at 50% 50%) - computed
FAIL test unit: em - ellipse(50em 100em) - computed assert_equals: expected "ellipse(800px 1600px)" but got "ellipse(800px 1600px at 50% 50%)"
FAIL test unit: em - ellipse(100em 100px) - computed assert_equals: expected "ellipse(1600px 100px)" but got "ellipse(1600px 100px at 50% 50%)"
FAIL test unit: em - ellipse(50em 25%) - computed assert_equals: expected "ellipse(800px 25%)" but got "ellipse(800px 25% at 50% 50%)"
FAIL test unit: em - ellipse(25% 50em) - computed assert_equals: expected "ellipse(25% 800px)" but got "ellipse(25% 800px at 50% 50%)"
FAIL test unit: em - ellipse(25em closest-side) - computed assert_equals: expected "ellipse(400px closest-side)" but got "ellipse(400px closest-side at 50% 50%)"
FAIL test unit: em - ellipse(closest-side 75em) - computed assert_equals: expected "ellipse(closest-side 1200px)" but got "ellipse(closest-side 1200px at 50% 50%)"
FAIL test unit: em - ellipse(25em farthest-side) - computed assert_equals: expected "ellipse(400px farthest-side)" but got "ellipse(400px farthest-side at 50% 50%)"
FAIL test unit: em - ellipse(farthest-side 75em) - computed assert_equals: expected "ellipse(farthest-side 1200px)" but got "ellipse(farthest-side 1200px at 50% 50%)"
PASS test unit: ex - ellipse(50ex 100ex at 50% 50%) - computed
PASS test unit: ex - ellipse(100ex 100px at 50% 50%) - computed
PASS test unit: ex - ellipse(50ex 25% at 50% 50%) - computed
PASS test unit: ex - ellipse(25% 50ex at 50% 50%) - computed
PASS test unit: ex - ellipse(25ex closest-side at 50% 50%) - computed
PASS test unit: ex - ellipse(closest-side 75ex at 50% 50%) - computed
PASS test unit: ex - ellipse(25ex farthest-side at 50% 50%) - computed
PASS test unit: ex - ellipse(farthest-side 75ex at 50% 50%) - computed
FAIL test unit: ex - ellipse(50ex 100ex) - computed assert_equals: expected "ellipse(640px 1280px)" but got "ellipse(640px 1280px at 50% 50%)"
FAIL test unit: ex - ellipse(100ex 100px) - computed assert_equals: expected "ellipse(1280px 100px)" but got "ellipse(1280px 100px at 50% 50%)"
FAIL test unit: ex - ellipse(50ex 25%) - computed assert_equals: expected "ellipse(640px 25%)" but got "ellipse(640px 25% at 50% 50%)"
FAIL test unit: ex - ellipse(25% 50ex) - computed assert_equals: expected "ellipse(25% 640px)" but got "ellipse(25% 640px at 50% 50%)"
FAIL test unit: ex - ellipse(25ex closest-side) - computed assert_equals: expected "ellipse(320px closest-side)" but got "ellipse(320px closest-side at 50% 50%)"
FAIL test unit: ex - ellipse(closest-side 75ex) - computed assert_equals: expected "ellipse(closest-side 960px)" but got "ellipse(closest-side 960px at 50% 50%)"
FAIL test unit: ex - ellipse(25ex farthest-side) - computed assert_equals: expected "ellipse(320px farthest-side)" but got "ellipse(320px farthest-side at 50% 50%)"
FAIL test unit: ex - ellipse(farthest-side 75ex) - computed assert_equals: expected "ellipse(farthest-side 960px)" but got "ellipse(farthest-side 960px at 50% 50%)"
PASS test unit: ch - ellipse(50ch 100ch at 50% 50%) - computed
PASS test unit: ch - ellipse(100ch 100px at 50% 50%) - computed
PASS test unit: ch - ellipse(50ch 25% at 50% 50%) - computed
PASS test unit: ch - ellipse(25% 50ch at 50% 50%) - computed
PASS test unit: ch - ellipse(25ch closest-side at 50% 50%) - computed
PASS test unit: ch - ellipse(closest-side 75ch at 50% 50%) - computed
PASS test unit: ch - ellipse(25ch farthest-side at 50% 50%) - computed
PASS test unit: ch - ellipse(farthest-side 75ch at 50% 50%) - computed
FAIL test unit: ch - ellipse(50ch 100ch) - computed assert_equals: expected "ellipse(800px 1600px)" but got "ellipse(800px 1600px at 50% 50%)"
FAIL test unit: ch - ellipse(100ch 100px) - computed assert_equals: expected "ellipse(1600px 100px)" but got "ellipse(1600px 100px at 50% 50%)"
FAIL test unit: ch - ellipse(50ch 25%) - computed assert_equals: expected "ellipse(800px 25%)" but got "ellipse(800px 25% at 50% 50%)"
FAIL test unit: ch - ellipse(25% 50ch) - computed assert_equals: expected "ellipse(25% 800px)" but got "ellipse(25% 800px at 50% 50%)"
FAIL test unit: ch - ellipse(25ch closest-side) - computed assert_equals: expected "ellipse(400px closest-side)" but got "ellipse(400px closest-side at 50% 50%)"
FAIL test unit: ch - ellipse(closest-side 75ch) - computed assert_equals: expected "ellipse(closest-side 1200px)" but got "ellipse(closest-side 1200px at 50% 50%)"
FAIL test unit: ch - ellipse(25ch farthest-side) - computed assert_equals: expected "ellipse(400px farthest-side)" but got "ellipse(400px farthest-side at 50% 50%)"
FAIL test unit: ch - ellipse(farthest-side 75ch) - computed assert_equals: expected "ellipse(farthest-side 1200px)" but got "ellipse(farthest-side 1200px at 50% 50%)"
PASS test unit: rem - ellipse(50rem 100rem at 50% 50%) - computed
PASS test unit: rem - ellipse(100rem 100px at 50% 50%) - computed
PASS test unit: rem - ellipse(50rem 25% at 50% 50%) - computed
PASS test unit: rem - ellipse(25% 50rem at 50% 50%) - computed
PASS test unit: rem - ellipse(25rem closest-side at 50% 50%) - computed
PASS test unit: rem - ellipse(closest-side 75rem at 50% 50%) - computed
PASS test unit: rem - ellipse(25rem farthest-side at 50% 50%) - computed
PASS test unit: rem - ellipse(farthest-side 75rem at 50% 50%) - computed
FAIL test unit: rem - ellipse(50rem 100rem) - computed assert_equals: expected "ellipse(800px 1600px)" but got "ellipse(800px 1600px at 50% 50%)"
FAIL test unit: rem - ellipse(100rem 100px) - computed assert_equals: expected "ellipse(1600px 100px)" but got "ellipse(1600px 100px at 50% 50%)"
FAIL test unit: rem - ellipse(50rem 25%) - computed assert_equals: expected "ellipse(800px 25%)" but got "ellipse(800px 25% at 50% 50%)"
FAIL test unit: rem - ellipse(25% 50rem) - computed assert_equals: expected "ellipse(25% 800px)" but got "ellipse(25% 800px at 50% 50%)"
FAIL test unit: rem - ellipse(25rem closest-side) - computed assert_equals: expected "ellipse(400px closest-side)" but got "ellipse(400px closest-side at 50% 50%)"
FAIL test unit: rem - ellipse(closest-side 75rem) - computed assert_equals: expected "ellipse(closest-side 1200px)" but got "ellipse(closest-side 1200px at 50% 50%)"
FAIL test unit: rem - ellipse(25rem farthest-side) - computed assert_equals: expected "ellipse(400px farthest-side)" but got "ellipse(400px farthest-side at 50% 50%)"
FAIL test unit: rem - ellipse(farthest-side 75rem) - computed assert_equals: expected "ellipse(farthest-side 1200px)" but got "ellipse(farthest-side 1200px at 50% 50%)"
PASS test unit: vw - ellipse(50vw 100vw at 50% 50%) - computed
PASS test unit: vw - ellipse(100vw 100px at 50% 50%) - computed
PASS test unit: vw - ellipse(50vw 25% at 50% 50%) - computed
PASS test unit: vw - ellipse(25% 50vw at 50% 50%) - computed
PASS test unit: vw - ellipse(25vw closest-side at 50% 50%) - computed
PASS test unit: vw - ellipse(closest-side 75vw at 50% 50%) - computed
PASS test unit: vw - ellipse(25vw farthest-side at 50% 50%) - computed
PASS test unit: vw - ellipse(farthest-side 75vw at 50% 50%) - computed
FAIL test unit: vw - ellipse(50vw 100vw) - computed assert_equals: expected "ellipse(400px 800px)" but got "ellipse(400px 800px at 50% 50%)"
FAIL test unit: vw - ellipse(100vw 100px) - computed assert_equals: expected "ellipse(800px 100px)" but got "ellipse(800px 100px at 50% 50%)"
FAIL test unit: vw - ellipse(50vw 25%) - computed assert_equals: expected "ellipse(400px 25%)" but got "ellipse(400px 25% at 50% 50%)"
FAIL test unit: vw - ellipse(25% 50vw) - computed assert_equals: expected "ellipse(25% 400px)" but got "ellipse(25% 400px at 50% 50%)"
FAIL test unit: vw - ellipse(25vw closest-side) - computed assert_equals: expected "ellipse(200px closest-side)" but got "ellipse(200px closest-side at 50% 50%)"
FAIL test unit: vw - ellipse(closest-side 75vw) - computed assert_equals: expected "ellipse(closest-side 600px)" but got "ellipse(closest-side 600px at 50% 50%)"
FAIL test unit: vw - ellipse(25vw farthest-side) - computed assert_equals: expected "ellipse(200px farthest-side)" but got "ellipse(200px farthest-side at 50% 50%)"
FAIL test unit: vw - ellipse(farthest-side 75vw) - computed assert_equals: expected "ellipse(farthest-side 600px)" but got "ellipse(farthest-side 600px at 50% 50%)"
PASS test unit: vh - ellipse(50vh 100vh at 50% 50%) - computed
PASS test unit: vh - ellipse(100vh 100px at 50% 50%) - computed
PASS test unit: vh - ellipse(50vh 25% at 50% 50%) - computed
PASS test unit: vh - ellipse(25% 50vh at 50% 50%) - computed
PASS test unit: vh - ellipse(25vh closest-side at 50% 50%) - computed
PASS test unit: vh - ellipse(closest-side 75vh at 50% 50%) - computed
PASS test unit: vh - ellipse(25vh farthest-side at 50% 50%) - computed
PASS test unit: vh - ellipse(farthest-side 75vh at 50% 50%) - computed
FAIL test unit: vh - ellipse(50vh 100vh) - computed assert_equals: expected "ellipse(300px 600px)" but got "ellipse(300px 600px at 50% 50%)"
FAIL test unit: vh - ellipse(100vh 100px) - computed assert_equals: expected "ellipse(600px 100px)" but got "ellipse(600px 100px at 50% 50%)"
FAIL test unit: vh - ellipse(50vh 25%) - computed assert_equals: expected "ellipse(300px 25%)" but got "ellipse(300px 25% at 50% 50%)"
FAIL test unit: vh - ellipse(25% 50vh) - computed assert_equals: expected "ellipse(25% 300px)" but got "ellipse(25% 300px at 50% 50%)"
FAIL test unit: vh - ellipse(25vh closest-side) - computed assert_equals: expected "ellipse(150px closest-side)" but got "ellipse(150px closest-side at 50% 50%)"
FAIL test unit: vh - ellipse(closest-side 75vh) - computed assert_equals: expected "ellipse(closest-side 450px)" but got "ellipse(closest-side 450px at 50% 50%)"
FAIL test unit: vh - ellipse(25vh farthest-side) - computed assert_equals: expected "ellipse(150px farthest-side)" but got "ellipse(150px farthest-side at 50% 50%)"
FAIL test unit: vh - ellipse(farthest-side 75vh) - computed assert_equals: expected "ellipse(farthest-side 450px)" but got "ellipse(farthest-side 450px at 50% 50%)"
PASS test unit: vmin - ellipse(50vmin 100vmin at 50% 50%) - computed
PASS test unit: vmin - ellipse(100vmin 100px at 50% 50%) - computed
PASS test unit: vmin - ellipse(50vmin 25% at 50% 50%) - computed
PASS test unit: vmin - ellipse(25% 50vmin at 50% 50%) - computed
PASS test unit: vmin - ellipse(25vmin closest-side at 50% 50%) - computed
PASS test unit: vmin - ellipse(closest-side 75vmin at 50% 50%) - computed
PASS test unit: vmin - ellipse(25vmin farthest-side at 50% 50%) - computed
PASS test unit: vmin - ellipse(farthest-side 75vmin at 50% 50%) - computed
FAIL test unit: vmin - ellipse(50vmin 100vmin) - computed assert_equals: expected "ellipse(300px 600px)" but got "ellipse(300px 600px at 50% 50%)"
FAIL test unit: vmin - ellipse(100vmin 100px) - computed assert_equals: expected "ellipse(600px 100px)" but got "ellipse(600px 100px at 50% 50%)"
FAIL test unit: vmin - ellipse(50vmin 25%) - computed assert_equals: expected "ellipse(300px 25%)" but got "ellipse(300px 25% at 50% 50%)"
FAIL test unit: vmin - ellipse(25% 50vmin) - computed assert_equals: expected "ellipse(25% 300px)" but got "ellipse(25% 300px at 50% 50%)"
FAIL test unit: vmin - ellipse(25vmin closest-side) - computed assert_equals: expected "ellipse(150px closest-side)" but got "ellipse(150px closest-side at 50% 50%)"
FAIL test unit: vmin - ellipse(closest-side 75vmin) - computed assert_equals: expected "ellipse(closest-side 450px)" but got "ellipse(closest-side 450px at 50% 50%)"
FAIL test unit: vmin - ellipse(25vmin farthest-side) - computed assert_equals: expected "ellipse(150px farthest-side)" but got "ellipse(150px farthest-side at 50% 50%)"
FAIL test unit: vmin - ellipse(farthest-side 75vmin) - computed assert_equals: expected "ellipse(farthest-side 450px)" but got "ellipse(farthest-side 450px at 50% 50%)"
PASS test unit: vmax - ellipse(50vmax 100vmax at 50% 50%) - computed
PASS test unit: vmax - ellipse(100vmax 100px at 50% 50%) - computed
PASS test unit: vmax - ellipse(50vmax 25% at 50% 50%) - computed
PASS test unit: vmax - ellipse(25% 50vmax at 50% 50%) - computed
PASS test unit: vmax - ellipse(25vmax closest-side at 50% 50%) - computed
PASS test unit: vmax - ellipse(closest-side 75vmax at 50% 50%) - computed
PASS test unit: vmax - ellipse(25vmax farthest-side at 50% 50%) - computed
PASS test unit: vmax - ellipse(farthest-side 75vmax at 50% 50%) - computed
FAIL test unit: vmax - ellipse(50vmax 100vmax) - computed assert_equals: expected "ellipse(400px 800px)" but got "ellipse(400px 800px at 50% 50%)"
FAIL test unit: vmax - ellipse(100vmax 100px) - computed assert_equals: expected "ellipse(800px 100px)" but got "ellipse(800px 100px at 50% 50%)"
FAIL test unit: vmax - ellipse(50vmax 25%) - computed assert_equals: expected "ellipse(400px 25%)" but got "ellipse(400px 25% at 50% 50%)"
FAIL test unit: vmax - ellipse(25% 50vmax) - computed assert_equals: expected "ellipse(25% 400px)" but got "ellipse(25% 400px at 50% 50%)"
FAIL test unit: vmax - ellipse(25vmax closest-side) - computed assert_equals: expected "ellipse(200px closest-side)" but got "ellipse(200px closest-side at 50% 50%)"
FAIL test unit: vmax - ellipse(closest-side 75vmax) - computed assert_equals: expected "ellipse(closest-side 600px)" but got "ellipse(closest-side 600px at 50% 50%)"
FAIL test unit: vmax - ellipse(25vmax farthest-side) - computed assert_equals: expected "ellipse(200px farthest-side)" but got "ellipse(200px farthest-side at 50% 50%)"
FAIL test unit: vmax - ellipse(farthest-side 75vmax) - computed assert_equals: expected "ellipse(farthest-side 600px)" but got "ellipse(farthest-side 600px at 50% 50%)"
